// document-models/scope-of-work/v1/gen/schema/types.ts
export type Maybe<T> = T | null | undefined;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Address: { input: `${string}:0x${string}`; output: `${string}:0x${string}` };
  Amount: {
    input: { unit?: string; value?: number };
    output: { unit?: string; value?: number };
  };
  Amount_Crypto: {
    input: { unit: string; value: string };
    output: { unit: string; value: string };
  };
  Amount_Currency: {
    input: { unit: string; value: string };
    output: { unit: string; value: string };
  };
  Amount_Fiat: {
    input: { unit: string; value: number };
    output: { unit: string; value: number };
  };
  Amount_Money: { input: number; output: number };
  Amount_Percentage: { input: number; output: number };
  Amount_Tokens: { input: number; output: number };
  AttachmentRef: {
    input: `attachment://v${number}:${string}`;
    output: `attachment://v${number}:${string}`;
  };
  Currency: { input: string; output: string };
  Date: { input: string; output: string };
  DateTime: { input: string; output: string };
  EmailAddress: { input: string; output: string };
  EthereumAddress: { input: string; output: string };
  OID: { input: string; output: string };
  OLabel: { input: string; output: string };
  PHID: { input: string; output: string };
  URL: { input: string; output: string };
  Unknown: { input: unknown; output: unknown };
  Upload: { input: File; output: File };
};
// ...
export type BudgetAnchorProject = {
  margin: Scalars["Float"]["output"];
  marginPinned: Maybe<Scalars["Boolean"]["output"]>;
  project: Maybe<Scalars["OID"]["output"]>;
  quantity: Scalars["Float"]["output"];
  unit: Maybe<Unit>;
  unitCost: Scalars["Float"]["output"];
};

export type BudgetExpenditure = {
  actuals: Scalars["Float"]["output"];
  cap: Scalars["Float"]["output"];
  percentage: Scalars["Float"]["output"];
};

export type BudgetType = "CAPEX" | "CONTINGENCY" | "OPEX" | "OVERHEAD";

export type Deliverable = {
  budgetAnchor: Maybe<BudgetAnchorProject>;
  code: Scalars["String"]["output"];
  description: Scalars["String"]["output"];
  goalRef: Maybe<Scalars["OID"]["output"]>;
  icon: Maybe<Scalars["String"]["output"]>;
  id: Scalars["OID"]["output"];
  keyResults: Array<KeyResult>;
  owner: Maybe<Scalars["ID"]["output"]>;
  status: DeliverableStatus;
  title: Scalars["String"]["output"];
  workProgress: Maybe<Progress>;
};

export type DeliverableSetStatus =
  | "CANCELED"
  | "DRAFT"
  | "FINISHED"
  | "IN_PROGRESS"
  | "TODO";

export type DeliverableSetStatusInput =
  | "CANCELED"
  | "DRAFT"
  | "FINISHED"
  | "IN_PROGRESS"
  | "TODO";

export type DeliverableStatus =
  | "BLOCKED"
  | "CANCELED"
  | "DELIVERED"
  | "DRAFT"
  | "IN_PROGRESS"
  | "TODO"
  | "WONT_DO";
// ...
export type KeyResult = {
  id: Scalars["OID"]["output"];
  link: Scalars["String"]["output"];
  title: Scalars["String"]["output"];
};
// ...
export type Progress = {
  completed: Maybe<Scalars["Int"]["output"]>;
  done: Maybe<Scalars["Boolean"]["output"]>;
  total: Maybe<Scalars["Int"]["output"]>;
  value: Maybe<Scalars["Float"]["output"]>;
};
// ...
export type Unit = "Hours" | "StoryPoints";